Fixes, updates, rearrangements
Fixed the code for fitting with gauss. Updated some of the docstrings. Added a function for plotting valid timestamps vs pixel using the calibration data. Added function for fitting with gauss using the calibration data. Moved Jakub's playground to _old. The app was moved to a separate repo to avoid having two mains and for an option to create an executable for the app. The code, however, will stay in this repo as well for the time being.