Major update
Updated both the app and the main library: now both the mask and the calibration data are searched for automatically by the app; the board number is selected in a combo box. Improved the path implementation: should work on any machine with no need to show path to the calibration or mask data, just to have it in place where the code will look for it.