Major update
Rewrote the memory-friendly version of the plot_grid function for plotting delta ts, as the previous version was crushing for over 10 GB of data. Renamed all the functions that utilize the calibration data as the ones that don't are now redundant (those are now ine the _old folder). Updated documentation. The new delta t function fig output is weird, to be solved.